Port "app-document" to e2e
