+++ test/e2e/app-document/csp.test.ts
@@ -1,0 +1,23 @@
+import { nextTestSetup } from 'e2e-utils'
+
+describe('With CSP enabled', () => {
+  const { next } = nextTestSetup({
+    files: __dirname,
+  })
+
+  it('should load inline script by hash', async () => {
+    const browser = await next.browser('/?withCSP=hash')
+    if (global.browserName === 'chrome') {
+      const errLog = await browser.log()
+      expect(errLog.filter((e) => e.source === 'security')).toEqual([])
+    }
+  })
+
+  it('should load inline script by nonce', async () => {
+    const browser = await next.browser('/?withCSP=nonce')
+    if (global.browserName === 'chrome') {
+      const errLog = await browser.log()
+      expect(errLog.filter((e) => e.source === 'security')).toEqual([])
+    }
+  })
+})
